+++ parsers/css.js
@@ -1,19 +1,4 @@
-export function createStyleSheet(rules, { media, disabled, baseURL } = {}) {
-	if (media instanceof MediaQueryList) {
-		return createStyleSheet(rules, { media: media.media, disabled, baseURL });
-	} else {
-		const sheet = new CSSStyleSheet({ media, disabled, baseURL });
-		sheet.replaceSync(rules);
-
-		return sheet;
-	}
-}
-
-export function createCSSParser({ media, disabled, baseURL } = {}) {
-	return (...args) => createStyleSheet(String.raw.apply(null, args), { media, disabled, baseURL });
-}
-
-export const css = createCSSParser({ baseURL: document.baseURI });
+import { createStyleSheet, createCSSParser, css, setStyleSheets, addStyleSheets } from '@aegisjsproject/parsers/css.js';
 
 export const lightCSS = createCSSParser({ media: '(prefers-color-scheme: light)', baseURL: document.baseURI });
 
@@ -43,3 +28,5 @@ export function styleSheetToLink(styleSheet) {
 
 	return link;
 }
+
+export { createStyleSheet, createCSSParser, css, setStyleSheets, addStyleSheets };