+++ parsers/svg.js
@@ -1,8 +1,1 @@
-import { elements, attributes } from '@aegisjsproject/sanitizer/config/svg.js';
-
-export function svg(strings, ...args) {
-	return Document.parseHTML(
-		String.raw(strings, ...args),
-		{ sanitizer: { elements: elements.concat(['html', 'body', 'head']), attributes }}
-	).body.firstElementChild;
-}
+export { svg } from '@aegisjsproject/parsers/svg.js';