+++ parsers/xml.js
@@ -1,4 +1,1 @@
-import { stringify } from '../stringify.js';
-
-export const xml = (strings, ...values) => new DOMParser()
-	.parseFromString(String.raw(strings, ...values.map(stringify)), 'application/xml');
+export { xml } from '@aegisjsproject/parsers/xml.js';